+++ app/src/main/java/woowacourse/payments/ui/cardlist/CardListContents.kt
@@ -25,7 +25,7 @@ import woowacourse.payments.domain.CardNumber
 import woowacourse.payments.domain.CardholderName
 import woowacourse.payments.domain.ExpirationDate
 import woowacourse.payments.domain.Passcode
-import woowacourse.payments.ui.addcard.PaymentCard
+import woowacourse.payments.ui.sharedcomposable.PaymentCard
 import woowacourse.payments.ui.model.toUiModel
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 import java.time.YearMonth
